Refactoring message pan gesture handling into its own class.
